Fixing init for Hospital
The way Hospital was initialized would cause issues with fetching. Whenever a fetch operation is called, CoreData assumes there to be an initializer called 'init(entity: NSEntityDescription, insertInto context: NSManagedObjectContext?)'. It is fair to assume that, since there is one in NSManagedObject. But the way the code was, NSManagedObject's initializers were not inherited.

By turning the init we have into a convenience initializer, we fix it to be what we had in mind: just a convenient way to initalize a Hospital with all its prop
erties.